Move `MultiHeadRelativeAttention` to relative_attention.py.
PiperOrigin-RevId: 331652519
